requirements.txt
fastapi
torch
transformers
python-dotenv
uvicorn